refactor: Move openapi tests to expected folder
