+++ packages/cli/tests/unit/openapi/util.ts
